feat(brain): in-process compactor scheduler (#961)
- Extract all six compaction jobs from scripts/compact-facts.ts into
  src/lib/cortex/compactor.ts (runCompaction + CompactionArgs + CompactionResult)
- Add src/lib/cortex/compactor-scheduler.ts: boots from ws-server,
  checks ~/.o8/last-compaction, fires once if >24h since last run,
  then repeats on a 24h interval; all timers unref so they don't
  block process exit; disable via O8_DISABLE_COMPACTOR=1
- Logs each run to ~/.o8/logs/compactor.log (append-only)
- Refactor scripts/compact-facts.ts to a thin CLI wrapper over the lib;
  external behavior is identical so the launchd plist keeps working
